vision-agent-service/pyproject.toml
[project]
name = "vision-agents-golf"
version = "0.0.0"
requires-python = ">=3.13"

# put only what this example needs
dependencies = [
  "python-dotenv>=1.0",
  "vision-agents",
  "vision-agents-plugins-openai",
  "vision-agents-plugins-getstream",
  "vision-agents-plugins-ultralytics",
  "vision-agents-plugins-gemini",
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
  "httpx[socks]>=0.25.0",
]

[tool.uv.sources]
"vision-agents" = {path = "./agents-core", editable=true}
"vision-agents-plugins-ultralytics" = {path = "./plugins/ultralytics", editable=true}
"vision-agents-plugins-getstream" = {path = "./plugins/getstream", editable=true}
"vision-agents-plugins-gemini" = {path = "./plugins/gemini", editable=true}